checkAuthorization()으로 변경하여 token 파싱을 내부에서 처리하고, token 통해 전송된 유저 검증 추가

+++ app/src/main/java/com/codesoom/assignment/controllers/ProductController.java
@@ -39,7 +39,7 @@ public class ProductController {
             @RequestHeader("Authorization") String authorization,
             @RequestBody @Valid ProductData productData
     ) {
-        authorizationService.parseToken(authorization);
+        authorizationService.checkUserAuthorization(authorization);
 
         return productService.createProduct(productData);
     }
@@ -50,7 +50,7 @@ public class ProductController {
             @PathVariable Long id,
             @RequestBody @Valid ProductData productData
     ) {
-        authorizationService.parseToken(authorization);
+        authorizationService.checkUserAuthorization(authorization);
 
         return productService.updateProduct(id, productData);
     }
@@ -61,7 +61,7 @@ public class ProductController {
             @RequestHeader("Authorization") String authorization,
             @PathVariable Long id
     ) {
-        authorizationService.parseToken(authorization);
+        authorizationService.checkUserAuthorization(authorization);
 
         productService.deleteProduct(id);
     }
